botEditor/backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
PyJWT
passlib[bcrypt]==1.7.4
bcrypt==3.2.2